requirements.txt
imageio~=2.34.1
emoji~=2.12.1
moviepy~=1.0.3
pillow~=10.3.0
requests~=2.32.3
cloudscraper~=1.2.71
tenacity~=8.4.1
beautifulsoup4~=4.12.3
typing_extensions~=4.12.2
pywebview~=5.1